refactor(time): refactor root barrel pattern to distinguish import type
