--- apps/time/src/widgets/home/index.ts
@@ -1,1 +1,0 @@
-export * from './ui';
