--- apps/time/src/widgets/menu/index.ts
@@ -1,1 +1,0 @@
-export * from './ui';
