--- apps/time/src/widgets/time-table/index.ts
@@ -1,4 +1,0 @@
-export * from './ui';
-export * from './model';
-export * from './types';
-export * from './api';
