+++ apps/time/src/widgets/time-table/ui/TimeTableContainer.tsx
@@ -2,7 +2,7 @@
 
 import { Suspense } from 'react';
 
-import { TimeTableLayout } from '@/widgets/time-table';
+import { TimeTableLayout } from '@/widgets/time-table/ui';
 
 export default function TimeTableContainer() {
   return (
